+++ examples/multiple/main.go
@@ -23,9 +23,9 @@ func main() {
 	fmt.Println("After Cursor", response.Cursor.After)
 	fmt.Println("Before Cursor", response.Cursor.Before)
 
-	for _, data := range response.Data {
-		fmt.Println("Block Hash:", data.Block.Hash)
-		fmt.Println("ID:", data.Data.ID)
-		fmt.Println("Code:", data.Data.Code)
+	for _, record := range response.Data {
+		fmt.Println("Block Hash:", record.Block.Hash)
+		fmt.Println("ID:", record.Data.ID)
+		fmt.Println("Code:", record.Data.Code)
 	}
 }